requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pytesseract
pdfplumber
Pillow
python-docx
openpyxl
pandas
openai
